requirements.txt
requests==2.32.3
pillow==11.2.1
python-dotenv==1.1.0